moving athena helpers to CLI where they are used
